renamed endpoint to endpoints
