+++ src/endpoints/slug-endpoints.js
