+++ src/endpoints/static-page-endpoints.js
